Squash of feature/up-linux-streamer-as-zenoh-plugin
Adding project to Cargo.toml

Noted area where we fail. Seems to be when initializing Session based on Runtime.

Had to lower the rustc version of up-rust to 1.72.0 and thus put the up-transport libs on test branches using that change, but now appears like it works :)

Able to use 1.74.0. Should now start to roll back my bespoke changes to allow us to use 1.72.0.

Rolled back to using stock branches. Was able to use Zenoh on its 0.11.0-rc3, but changing its rust-toolchain.toml to bump to 1.74.0 so that we have compatability between zenohd and the up-linux-streamer-plugin

Addind documentation on how to build and use the up-linux-streamer-plugin with the own-compiled version of zenoh

Add notes on where issues are at

Several more attempts to try to read the Config from the Runtime